apiary/requirements.txt
asgiref==3.4.1
attrs==21.2.0
black==21.9b0
click==8.0.1
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
drf-spectacular==0.13.2
factory-boy==3.2.0
Faker==8.14.0
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.0.1
Markdown==3.3.4
mypy-extensions==0.4.3
mysqlclient==2.0.3
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-django==4.4.0
pytest-mock==3.6.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.9.24
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
